pyproject.toml
[project]
name = "boursobank-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "glops", email = "skylvl@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.8.2",
    "keyring>=25.6.0",
    "msgspec>=0.19.0",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
boursobank-scraper = "boursobank_scraper.__main__:main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
